refactor: extract `check_assertions`
